share upload results between redundant upload-jobs (#454)
In cases where multiple upload/processing-jobs refer to the same
push-target, thus-redundant upload-jobs were discarded too early. This
resulting in those jobs to no longer have means to retrieve uploaded
oci-manifests (or rather: their hash-digests).

Move deduplication logic to a later point in time. Mark an immanent
uploading operation in a dict (and guard the operation w/ an event). Any
subsequent uploader will wait for this event to be set (marking the
upload as done). At which point in time the resulting manifest-digest is
available in a second dict.