workaround: ignore errors in schema-ref-resolution
For reasons not clearly understood, `jsonschema` lib seemingly tried to
resolve some external references from component-descriptor's
json-schema.

As there are no external references in used schema (except the
json-schema-id, which is built into jsonschema module, and does not lead
to external http-requests), it is not entirely clear what caused
those requests. Unfortunately, jsonschema will not log the target-URLs.

From error-message, it is clear that some error-document was returned,
instead of the expedted JSON-body. There are, however, no
status-code-checks.

As a workaround, to prevent similar errors from happening again,
downgrade the execption raised upon reference-resolution errors to just
a warning to at least not block lssd-releases.